+++ lib/run.js
@@ -1,9 +1,10 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -13,7 +14,7 @@ const core = require("@actions/core");
 const io = require("@actions/io");
 const path = require("path");
 const utility_1 = require("./utilities/utility");
-const kubectl_util_1 = require("./kubectl-util");
+const kubectl_util_1 = require("./utilities/kubectl-util");
 const deployment_helper_1 = require("./utilities/strategy-helpers/deployment-helper");
 const promote_1 = require("./actions/promote");
 const reject_1 = require("./actions/reject");
