+++ lib/utilities/strategy-helpers/smi-canary-deployment-helper.js
@@ -8,12 +8,12 @@ const TaskInputParameters = require("../../input-parameters");
 const fileHelper = require("../files-helper");
 const helper = require("../resource-object-utility");
 const utils = require("../manifest-utilities");
-const kubectlUtils = require("../../kubectl-util");
+const kubectlUtils = require("../kubectl-util");
 const canaryDeploymentHelper = require("./canary-deployment-helper");
 const utility_1 = require("../utility");
 const TRAFFIC_SPLIT_OBJECT_NAME_SUFFIX = '-workflow-rollout';
 const TRAFFIC_SPLIT_OBJECT = 'TrafficSplit';
-var trafficSplitAPIVersion = null;
+let trafficSplitAPIVersion = "";
 function deploySMICanary(kubectl, filePaths) {
     const newObjectsList = [];
     const canaryReplicaCount = parseInt(TaskInputParameters.baselineAndCanaryReplicas);
